refactor(platform): add github and admin subdomains for v1 routes
