+++ backend/src/domains/platform/admin/index.ts
@@ -1,0 +1,2 @@
+export { default as AdminV1Controller } from './admin.v1.controller';
+export { default as adminV1Router } from './admin.v1.routes';
