+++ backend/src/domains/platform/github/github.types.ts
@@ -1,0 +1,10 @@
+/**
+ * Shape of a contributor entry as returned by the GitHub REST API.
+ */
+export interface GitHubContributor {
+  login: string;
+  type: string;
+  avatar_url: string;
+  contributions: number;
+  html_url: string;
+}
