+++ backend/src/domains/platform/github/index.ts
@@ -1,0 +1,5 @@
+export { default as GithubService } from './github.service';
+export { default as GithubController } from './github.controller';
+export { default as GithubGateway } from './github.gateway';
+export { default as githubV1Router } from './github.v1.routes';
+export type { GitHubContributor } from './github.types';
